@@ -1,15 +1,15 @@
 name: dart_appwrite
-version: 0.4.0
+version: 0.5.0-dev.1
 description: Appwrite is an open-source self-hosted backend server that abstract and simplify complex and repetitive development tasks behind a very simple REST API
 homepage: https://appwrite.io
 repository: https://github.com/appwrite/sdk-for-dart
 issue_tracker: https://github.com/appwrite/sdk-generator/issues
 documentation: https://appwrite.io/support
 environment:
-  sdk: '>=2.6.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 dependencies:
-  dio: ^3.0.10
-  meta: ^1.1.8
+  dio: ^4.0.0
+  meta: ^1.3.0
 
 dev_dependencies:
-  test:
+  test: ^1.16.8
