@@ -3,22 +3,23 @@ part of dart_appwrite;
 class Client {
     String endPoint;
     String type = 'unknown';
-    Map<String, String> headers;
-    Map<String, String> config;
+    Map<String, String>? headers;
+    late Map<String, String> config;
     bool selfSigned;
     bool initialized = false;
     Dio http;
 
-    Client({this.endPoint = 'https://appwrite.io/v1', this.selfSigned = false, Dio http}) : this.http = http ?? Dio() {
+    Client({this.endPoint = 'https://appwrite.io/v1', this.selfSigned = false, Dio? http}) : this.http = http ?? Dio() {
         
         this.headers = {
             'content-type': 'application/json',
-            'x-sdk-version': 'appwrite:dart:0.4.0',
+            'x-sdk-version': 'appwrite:dart:0.5.0-dev.1',
         };
 
         this.config = {};
 
         assert(endPoint.startsWith(RegExp("http://|https://")), "endPoint $endPoint must start with 'http'");
+        init();
     }
 
 
@@ -54,19 +55,17 @@ class Client {
     }
 
     Client addHeader(String key, String value) {
-        headers[key] = value;
+        headers![key] = value;
         
         return this;
     }
 
-    Future init() async {
-        if(!initialized) {
+    void init() {
           this.http.options.baseUrl = this.endPoint;
-          this.http.options.validateStatus = (status) => status < 400;
-        }
+          this.http.options.validateStatus = (status) => status! < 400;
     }
 
-    Future<Response> call(HttpMethod method, {String path = '', Map<String, String> headers = const {}, Map<String, dynamic> params = const {}, ResponseType responseType}) async {
+    Future<Response> call(HttpMethod method, {String path = '', Map<String, String> headers = const {}, Map<String, dynamic> params = const {}, ResponseType? responseType}) async {
         if(selfSigned) {
             // Allow self signed requests
             (http.httpClientAdapter as DefaultHttpClientAdapter).onHttpClientCreate = (HttpClient client) {
@@ -75,18 +74,17 @@ class Client {
             };
         }
 
-        await this.init();
-
         // Origin is hardcoded for testing
         Options options = Options(
-            headers: {...this.headers, ...headers},
+            headers: {...this.headers!, ...headers},
             method: method.name(),
-            responseType: responseType
+            responseType: responseType,
+            listFormat: ListFormat.multiCompatible,
         );
         try {
 
             if(headers['content-type'] == 'multipart/form-data') {
-                return await http.request(path, data: FormData.fromMap(params), options: options);
+                return await http.request(path, data: FormData.fromMap(params,ListFormat.multiCompatible), options: options);
             }
 
             if (method == HttpMethod.get) {
@@ -103,16 +101,16 @@ class Client {
             throw AppwriteException(e.message);
           }
           if(responseType == ResponseType.bytes) {
-            if(e.response.headers['content-type'].contains('application/json')) {
-              final res = json.decode(utf8.decode(e.response.data));
-              throw AppwriteException(res['message'],res['code'], e.response);
+            if(e.response!.headers['content-type']?.contains('application/json') ?? false) {
+              final res = json.decode(utf8.decode(e.response!.data));
+              throw AppwriteException(res['message'],res['code'], res);
             } else {
               throw AppwriteException(e.message);
             }
           }
-          throw AppwriteException(e.response.data['message'],e.response.data['code'], e.response.data);
+          throw AppwriteException(e.response?.data['message'],e.response?.data['code'], e.response?.data);
         } catch(e) {
-          throw AppwriteException(e.message);
+          throw AppwriteException(e.toString());
         }
     }
 }
