@@ -1,8 +1,8 @@
 part of dart_appwrite;
 
 class AppwriteException implements Exception {
-  final String message;
-  final int code;
+  final String? message;
+  final int? code;
   final dynamic response;
 
   AppwriteException([this.message = "", this.code, this.response]);
