@@ -11,7 +11,7 @@ class Avatars extends Service {
      /// /account/sessions endpoint. Use width, height and quality arguments to
      /// change the output settings.
      ///
-    Future<Response> getBrowser({@required String code, int width = 100, int height = 100, int quality = 100}) {
+    Future<Response> getBrowser({required String code, int width = 100, int height = 100, int quality = 100}) {
         final String path = '/avatars/browsers/{code}'.replaceAll(RegExp('{code}'), code);
 
         final Map<String, dynamic> params = {
@@ -35,7 +35,7 @@ class Avatars extends Service {
      /// provider you need. Use width, height and quality arguments to change the
      /// output settings.
      ///
-    Future<Response> getCreditCard({@required String code, int width = 100, int height = 100, int quality = 100}) {
+    Future<Response> getCreditCard({required String code, int width = 100, int height = 100, int quality = 100}) {
         final String path = '/avatars/credit-cards/{code}'.replaceAll(RegExp('{code}'), code);
 
         final Map<String, dynamic> params = {
@@ -59,7 +59,7 @@ class Avatars extends Service {
      /// website URL.
      /// 
      ///
-    Future<Response> getFavicon({@required String url}) {
+    Future<Response> getFavicon({required String url}) {
         final String path = '/avatars/favicon';
 
         final Map<String, dynamic> params = {
@@ -81,7 +81,7 @@ class Avatars extends Service {
      /// users. The code argument receives the 2 letter country code. Use width,
      /// height and quality arguments to change the output settings.
      ///
-    Future<Response> getFlag({@required String code, int width = 100, int height = 100, int quality = 100}) {
+    Future<Response> getFlag({required String code, int width = 100, int height = 100, int quality = 100}) {
         final String path = '/avatars/flags/{code}'.replaceAll(RegExp('{code}'), code);
 
         final Map<String, dynamic> params = {
@@ -106,7 +106,7 @@ class Avatars extends Service {
      /// remote images in your app or in case you want to make sure a 3rd party
      /// image is properly served using a TLS protocol.
      ///
-    Future<Response> getImage({@required String url, int width = 400, int height = 400}) {
+    Future<Response> getImage({required String url, int width = 400, int height = 400}) {
         final String path = '/avatars/image';
 
         final Map<String, dynamic> params = {
@@ -162,7 +162,7 @@ class Avatars extends Service {
      /// Converts a given plain text to a QR code image. You can use the query
      /// parameters to change the size and style of the resulting image.
      ///
-    Future<Response> getQR({@required String text, int size = 400, int margin = 1, bool download = false}) {
+    Future<Response> getQR({required String text, int size = 400, int margin = 1, bool download = false}) {
         final String path = '/avatars/qr';
 
         final Map<String, dynamic> params = {
