@@ -32,7 +32,7 @@ class Database extends Service {
      ///
      /// Create a new Collection.
      ///
-    Future<Response> createCollection({@required String name, @required List read, @required List write, @required List rules}) {
+    Future<Response> createCollection({required String name, required List read, required List write, required List rules}) {
         final String path = '/database/collections';
 
         final Map<String, dynamic> params = {
@@ -54,7 +54,7 @@ class Database extends Service {
      /// Get a collection by its unique ID. This endpoint response returns a JSON
      /// object with the collection metadata.
      ///
-    Future<Response> getCollection({@required String collectionId}) {
+    Future<Response> getCollection({required String collectionId}) {
         final String path = '/database/collections/{collectionId}'.replaceAll(RegExp('{collectionId}'), collectionId);
 
         final Map<String, dynamic> params = {
@@ -71,7 +71,7 @@ class Database extends Service {
      ///
      /// Update a collection by its unique ID.
      ///
-    Future<Response> updateCollection({@required String collectionId, @required String name, @required List read, @required List write, List rules = const []}) {
+    Future<Response> updateCollection({required String collectionId, required String name, required List read, required List write, List rules = const []}) {
         final String path = '/database/collections/{collectionId}'.replaceAll(RegExp('{collectionId}'), collectionId);
 
         final Map<String, dynamic> params = {
@@ -93,7 +93,7 @@ class Database extends Service {
      /// Delete a collection by its unique ID. Only users with write permissions
      /// have access to delete this resource.
      ///
-    Future<Response> deleteCollection({@required String collectionId}) {
+    Future<Response> deleteCollection({required String collectionId}) {
         final String path = '/database/collections/{collectionId}'.replaceAll(RegExp('{collectionId}'), collectionId);
 
         final Map<String, dynamic> params = {
@@ -113,7 +113,7 @@ class Database extends Service {
      /// of the project's documents. [Learn more about different API
      /// modes](/docs/admin).
      ///
-    Future<Response> listDocuments({@required String collectionId, List filters = const [], int limit = 25, int offset = 0, String orderField = '', OrderType orderType = OrderType.asc, String orderCast = 'string', String search = ''}) {
+    Future<Response> listDocuments({required String collectionId, List filters = const [], int limit = 25, int offset = 0, String orderField = '', OrderType orderType = OrderType.asc, String orderCast = 'string', String search = ''}) {
         final String path = '/database/collections/{collectionId}/documents'.replaceAll(RegExp('{collectionId}'), collectionId);
 
         final Map<String, dynamic> params = {
@@ -140,7 +140,7 @@ class Database extends Service {
      /// integration](/docs/server/database#databaseCreateCollection) API or
      /// directly from your database console.
      ///
-    Future<Response> createDocument({@required String collectionId, @required Map data, @required List read, @required List write, String parentDocument = '', String parentProperty = '', String parentPropertyType = 'assign'}) {
+    Future<Response> createDocument({required String collectionId, required Map data, required List read, required List write, String parentDocument = '', String parentProperty = '', String parentPropertyType = 'assign'}) {
         final String path = '/database/collections/{collectionId}/documents'.replaceAll(RegExp('{collectionId}'), collectionId);
 
         final Map<String, dynamic> params = {
@@ -164,7 +164,7 @@ class Database extends Service {
      /// Get a document by its unique ID. This endpoint response returns a JSON
      /// object with the document data.
      ///
-    Future<Response> getDocument({@required String collectionId, @required String documentId}) {
+    Future<Response> getDocument({required String collectionId, required String documentId}) {
         final String path = '/database/collections/{collectionId}/documents/{documentId}'.replaceAll(RegExp('{collectionId}'), collectionId).replaceAll(RegExp('{documentId}'), documentId);
 
         final Map<String, dynamic> params = {
@@ -182,7 +182,7 @@ class Database extends Service {
      /// Update a document by its unique ID. Using the patch method you can pass
      /// only specific fields that will get updated.
      ///
-    Future<Response> updateDocument({@required String collectionId, @required String documentId, @required Map data, @required List read, @required List write}) {
+    Future<Response> updateDocument({required String collectionId, required String documentId, required Map data, required List read, required List write}) {
         final String path = '/database/collections/{collectionId}/documents/{documentId}'.replaceAll(RegExp('{collectionId}'), collectionId).replaceAll(RegExp('{documentId}'), documentId);
 
         final Map<String, dynamic> params = {
@@ -204,7 +204,7 @@ class Database extends Service {
      /// documents, its attributes and relations to other documents. Child documents
      /// **will not** be deleted.
      ///
-    Future<Response> deleteDocument({@required String collectionId, @required String documentId}) {
+    Future<Response> deleteDocument({required String collectionId, required String documentId}) {
         final String path = '/database/collections/{collectionId}/documents/{documentId}'.replaceAll(RegExp('{collectionId}'), collectionId).replaceAll(RegExp('{documentId}'), documentId);
 
         final Map<String, dynamic> params = {
