@@ -32,7 +32,7 @@ class Functions extends Service {
      /// [permissions](/docs/permissions) to allow different project users or team
      /// with access to execute the function using the client API.
      ///
-    Future<Response> create({@required String name, @required List execute, @required String env, Map vars = const {}, List events = const [], String schedule = '', int timeout = 15}) {
+    Future<Response> create({required String name, required List execute, required String env, Map vars = const {}, List events = const [], String schedule = '', int timeout = 15}) {
         final String path = '/functions';
 
         final Map<String, dynamic> params = {
@@ -56,7 +56,7 @@ class Functions extends Service {
      ///
      /// Get a function by its unique ID.
      ///
-    Future<Response> get({@required String functionId}) {
+    Future<Response> get({required String functionId}) {
         final String path = '/functions/{functionId}'.replaceAll(RegExp('{functionId}'), functionId);
 
         final Map<String, dynamic> params = {
@@ -73,7 +73,7 @@ class Functions extends Service {
      ///
      /// Update function by its unique ID.
      ///
-    Future<Response> update({@required String functionId, @required String name, @required List execute, Map vars = const {}, List events = const [], String schedule = '', int timeout = 15}) {
+    Future<Response> update({required String functionId, required String name, required List execute, Map vars = const {}, List events = const [], String schedule = '', int timeout = 15}) {
         final String path = '/functions/{functionId}'.replaceAll(RegExp('{functionId}'), functionId);
 
         final Map<String, dynamic> params = {
@@ -96,7 +96,7 @@ class Functions extends Service {
      ///
      /// Delete a function by its unique ID.
      ///
-    Future<Response> delete({@required String functionId}) {
+    Future<Response> delete({required String functionId}) {
         final String path = '/functions/{functionId}'.replaceAll(RegExp('{functionId}'), functionId);
 
         final Map<String, dynamic> params = {
@@ -116,7 +116,7 @@ class Functions extends Service {
      /// return a list of all of the project's teams. [Learn more about different
      /// API modes](/docs/admin).
      ///
-    Future<Response> listExecutions({@required String functionId, String search = '', int limit = 25, int offset = 0, OrderType orderType = OrderType.asc}) {
+    Future<Response> listExecutions({required String functionId, String search = '', int limit = 25, int offset = 0, OrderType orderType = OrderType.asc}) {
         final String path = '/functions/{functionId}/executions'.replaceAll(RegExp('{functionId}'), functionId);
 
         final Map<String, dynamic> params = {
@@ -140,7 +140,7 @@ class Functions extends Service {
      /// updates on the current execution status. Once this endpoint is called, your
      /// function execution process will start asynchronously.
      ///
-    Future<Response> createExecution({@required String functionId}) {
+    Future<Response> createExecution({required String functionId}) {
         final String path = '/functions/{functionId}/executions'.replaceAll(RegExp('{functionId}'), functionId);
 
         final Map<String, dynamic> params = {
@@ -157,7 +157,7 @@ class Functions extends Service {
      ///
      /// Get a function execution log by its unique ID.
      ///
-    Future<Response> getExecution({@required String functionId, @required String executionId}) {
+    Future<Response> getExecution({required String functionId, required String executionId}) {
         final String path = '/functions/{functionId}/executions/{executionId}'.replaceAll(RegExp('{functionId}'), functionId).replaceAll(RegExp('{executionId}'), executionId);
 
         final Map<String, dynamic> params = {
@@ -176,7 +176,7 @@ class Functions extends Service {
      /// endpoint to switch the code tag that should be executed by the execution
      /// endpoint.
      ///
-    Future<Response> updateTag({@required String functionId, @required String tag}) {
+    Future<Response> updateTag({required String functionId, required String tag}) {
         final String path = '/functions/{functionId}/tag'.replaceAll(RegExp('{functionId}'), functionId);
 
         final Map<String, dynamic> params = {
@@ -195,7 +195,7 @@ class Functions extends Service {
      /// Get a list of all the project's code tags. You can use the query params to
      /// filter your results.
      ///
-    Future<Response> listTags({@required String functionId, String search = '', int limit = 25, int offset = 0, OrderType orderType = OrderType.asc}) {
+    Future<Response> listTags({required String functionId, String search = '', int limit = 25, int offset = 0, OrderType orderType = OrderType.asc}) {
         final String path = '/functions/{functionId}/tags'.replaceAll(RegExp('{functionId}'), functionId);
 
         final Map<String, dynamic> params = {
@@ -225,7 +225,7 @@ class Functions extends Service {
      /// 
      /// Use the "command" param to set the entry point used to execute your code.
      ///
-    Future<Response> createTag({@required String functionId, @required String command, @required MultipartFile code}) {
+    Future<Response> createTag({required String functionId, required String command, required MultipartFile code}) {
         final String path = '/functions/{functionId}/tags'.replaceAll(RegExp('{functionId}'), functionId);
 
         final Map<String, dynamic> params = {
@@ -244,7 +244,7 @@ class Functions extends Service {
      ///
      /// Get a code tag by its unique ID.
      ///
-    Future<Response> getTag({@required String functionId, @required String tagId}) {
+    Future<Response> getTag({required String functionId, required String tagId}) {
         final String path = '/functions/{functionId}/tags/{tagId}'.replaceAll(RegExp('{functionId}'), functionId).replaceAll(RegExp('{tagId}'), tagId);
 
         final Map<String, dynamic> params = {
@@ -261,7 +261,7 @@ class Functions extends Service {
      ///
      /// Delete a code tag by its unique ID.
      ///
-    Future<Response> deleteTag({@required String functionId, @required String tagId}) {
+    Future<Response> deleteTag({required String functionId, required String tagId}) {
         final String path = '/functions/{functionId}/tags/{tagId}'.replaceAll(RegExp('{functionId}'), functionId).replaceAll(RegExp('{tagId}'), tagId);
 
         final Map<String, dynamic> params = {
