@@ -35,7 +35,7 @@ class Teams extends Service {
      /// who will be able add new owners and update or delete the team from your
      /// project.
      ///
-    Future<Response> create({@required String name, List roles = const ["owner"]}) {
+    Future<Response> create({required String name, List roles = const ["owner"]}) {
         final String path = '/teams';
 
         final Map<String, dynamic> params = {
@@ -55,7 +55,7 @@ class Teams extends Service {
      /// Get a team by its unique ID. All team members have read access for this
      /// resource.
      ///
-    Future<Response> get({@required String teamId}) {
+    Future<Response> get({required String teamId}) {
         final String path = '/teams/{teamId}'.replaceAll(RegExp('{teamId}'), teamId);
 
         final Map<String, dynamic> params = {
@@ -73,7 +73,7 @@ class Teams extends Service {
      /// Update a team by its unique ID. Only team owners have write access for this
      /// resource.
      ///
-    Future<Response> update({@required String teamId, @required String name}) {
+    Future<Response> update({required String teamId, required String name}) {
         final String path = '/teams/{teamId}'.replaceAll(RegExp('{teamId}'), teamId);
 
         final Map<String, dynamic> params = {
@@ -92,7 +92,7 @@ class Teams extends Service {
      /// Delete a team by its unique ID. Only team owners have write access for this
      /// resource.
      ///
-    Future<Response> delete({@required String teamId}) {
+    Future<Response> delete({required String teamId}) {
         final String path = '/teams/{teamId}'.replaceAll(RegExp('{teamId}'), teamId);
 
         final Map<String, dynamic> params = {
@@ -110,7 +110,7 @@ class Teams extends Service {
      /// Get a team members by the team unique ID. All team members have read access
      /// for this list of resources.
      ///
-    Future<Response> getMemberships({@required String teamId, String search = '', int limit = 25, int offset = 0, OrderType orderType = OrderType.asc}) {
+    Future<Response> getMemberships({required String teamId, String search = '', int limit = 25, int offset = 0, OrderType orderType = OrderType.asc}) {
         final String path = '/teams/{teamId}/memberships'.replaceAll(RegExp('{teamId}'), teamId);
 
         final Map<String, dynamic> params = {
@@ -143,7 +143,7 @@ class Teams extends Service {
      /// the only valid redirect URL's are the once from domains you have set when
      /// added your platforms in the console interface.
      ///
-    Future<Response> createMembership({@required String teamId, @required String email, @required List roles, @required String url, String name = ''}) {
+    Future<Response> createMembership({required String teamId, required String email, required List roles, required String url, String name = ''}) {
         final String path = '/teams/{teamId}/memberships'.replaceAll(RegExp('{teamId}'), teamId);
 
         final Map<String, dynamic> params = {
@@ -166,7 +166,7 @@ class Teams extends Service {
      /// the membership of any other team member. You can also use this endpoint to
      /// delete a user membership even if it is not accepted.
      ///
-    Future<Response> deleteMembership({@required String teamId, @required String inviteId}) {
+    Future<Response> deleteMembership({required String teamId, required String inviteId}) {
         final String path = '/teams/{teamId}/memberships/{inviteId}'.replaceAll(RegExp('{teamId}'), teamId).replaceAll(RegExp('{inviteId}'), inviteId);
 
         final Map<String, dynamic> params = {
